refactoring: move loggable message to logging namespace
